requirements.txt
imageio==2.3.0
matplotlib==2.2.2
numpy==1.15.0
pandas==0.23.4
Pillow==5.2.0
PyYAML==5.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
torch==1.4.0
torchvision==0.2.1
tqdm==4.24.0
